Refactored into single test file
